@@ -34,13 +34,13 @@ func (p *FilePermissions) UnmarshalYAML(value *yaml.Node) error {
 	var strValue string
 	err = value.Decode(&strValue)
 	if err != nil {
-		return fmt.Errorf("failed to parse FilePermissions: %w", err)
+		return fmt.Errorf("failed to parse FilePermissions:\n%w", err)
 	}
 
 	// Try to parse the string as an octal number.
 	fileModeUint, err := strconv.ParseUint(strValue, 8, 32)
 	if err != nil {
-		return fmt.Errorf("failed to parse FilePermissions: %w", err)
+		return fmt.Errorf("failed to parse FilePermissions:\n%w", err)
 	}
 
 	*p = (FilePermissions)(fileModeUint)
